cart/requirements.txt
# Fast API
fastapi
uvicorn

# Config
python-dotenv

# V1
redis
